Few HTML tag issues
Hello!

I want to report that there seems to be a couple of issues with HTML closing tags.
First, it seems to be missing a closing tag at line 667. The div opens at line 548, but never closes.
Second, it appears that there is an incorrect or misplaced closing tag at line 741.

Can you confirm it?

Thanks!