Frontend HTML improvements. Style loading improvements.
Removes the wrapper around the content and instead adds a div above the content, which the JS targets, gets its parent, and then initializes the child elements for the share events. 

Critical CSS like tooltips and highlights are inlined in the header for performance.

CSS that affects the icon appearance is loaded in the footer. Tooltips and highlights are only shown on hover, so it's not critical the CSS load first in the header.

The HTML improvements will provide better compatibility, and the style loading improvements should increase performance significantly.